backend/requirements-dev.txt
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
